requirements.txt
requests==2.29.0
fastapi~=0.95.1
uvicorn==0.22.0
gql[all]
python-dotenv